+++ js/const.js
@@ -1,0 +1,16 @@
+const HOTELS_COUNT = 10;
+const LAT_START = 35.65;
+const LAT_END = 35.7;
+const LNG_START = 139.7;
+const LNG_END = 139.8;
+const BUILDINGS = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
+const CHECK_TIME = ['12:00', '13:00', '14:00'];
+const TYPE_OF_ROOM = ['standart', 'comfort', 'luxe', 'president'];
+const FACILITIES = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
+const PHOTO_EXAMPLES = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg'
+];
+
+export { HOTELS_COUNT, LAT_START, LAT_END, LNG_START, LNG_END, BUILDINGS, CHECK_TIME, TYPE_OF_ROOM, FACILITIES, PHOTO_EXAMPLES };
