+++ js/utils/shuffle-arr.js
@@ -1,16 +1,5 @@
 import { returnRandomDiceNum } from './return-random-dice-num.js';
 
-// Функия для генерации массива случайной длинны из значений передаваемого массива (элементы повторяются)
-export const growArr = (arr) => {
-  const newArr = [];
-
-  for (let index = 0; index <= returnRandomDiceNum(1, 10); index++) {
-    newArr[index] = arr[returnRandomDiceNum(0, arr.length - 1)];
-  }
-
-  return newArr;
-};
-
 // Функия для генерации массива случайной длинны из значений передаваемого массива (элементы НЕ повторяются)
 export const shuffleArr = (arr) => {
   const newArr = [...arr];
